Factor Spectroscopy game into its own file
